Review comments
1. Move _response_forbidden into subs browser
2. Remove constants import from embed_extractor
