--- resources/lib/__init__.py
@@ -1,4 +1,0 @@
-
-
-
-
